Sign the agreement to receive the prize money.
.Sign the agreement to receive the prize money. And acknowledge the conditions in the different details Must inform you that my startup was due to wanting to earn extra Financially to see and family and to use some money to improve to help facilitate the development of work that is more effective Due to experience financial problems Therefore see the way
And keep going until this point of change And sometimes confusing and confused whether it is true or lie I am Thai who read Thai books, not 8 lines a day. But because I have a need And curious about the various facts I have to spend a lot of time by coppy every letter in order to translate. Understand and come to this point That means I pay attention and read it.
More than 8 lines Please be informed of my efforts and intentions.